+++ Plugins/CoreAPI/stories/mocks/index.js
@@ -1,0 +1,4 @@
+import get_customers_mocks from "./get-customers.js";
+import get_project_summaries from "./get-project-summaries.js";
+
+export default [...get_customers_mocks, ...get_project_summaries];
