+++ Plugins/CoreAPI/stories/setup-stories.stories.js
@@ -1,0 +1,4 @@
+import { registerStorybookMocks } from "meteor/penpal";
+import mocks from "./mocks";
+
+registerStorybookMocks("CoreAPI", mocks);
