+++ Server/src/package.json
@@ -30,7 +30,7 @@
         "classnames": "^2.2.6",
         "clipboard": "^2.0.6",
         "dataloader": "^2.0.0",
-        "faker": "^4.1.0",
+        "faker": "^5.3.1",
         "file-saver": "^2.0.2",
         "glob": "^7.1.6",
         "graphql": "^15.0.0",
