+++ Server/src/stories/login-page.stories.js
@@ -3,8 +3,8 @@ import _ from "lodash";
 import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
-import { Components } from "../client/modules/components.js";
-import { SetupProviders } from "./common.js";
+import { Components } from "meteor/penpal";
+import { SetupProviders } from "stories/common.js";
 
 const explore = storiesOf("PenPal/Login", module);
