+++ Server/src/stories/styled-components.stories.js
@@ -2,8 +2,8 @@ import React, { useState } from "react";
 import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
-import { Components } from "../client/modules/components.js";
-import { SetupProviders } from "./common.js";
+import { Components } from "meteor/penpal";
+import { SetupProviders } from "stories/common.js";
 
 const projects = storiesOf("UI/Styled Components", module);
 
@@ -52,7 +52,7 @@ const items = [
 
 projects.add("Select", () => {
   const [selected, setSelected] = useState("");
-  const handleChange = event => setSelected(event.target.value);
+  const handleChange = (event) => setSelected(event.target.value);
 
   return (
     <SetupProviders>
