+++ docker-compose.dev.yaml
@@ -33,6 +33,8 @@ services:
             - ./Server/src/.storybook:/app/.storybook
             - ./Server/src/lib/:/app/lib:consistent
             - ./Server/src/public:/app/public:consistent
+            - ./Server/src/packages:/app/packages:consistent
+            - ./Plugins/:/app/plugins:consistent
             - ./Server/src/.babelrc:/app/.babelrc:consistent
         networks:
             - public
